:zap: 将数据模型中一些 `isize` / `usize` 替换为更精确的数值类型

+++ crates/webgal-ls/src/project/ident.rs
@@ -12,7 +12,7 @@ pub struct IdentTable {
     pub speaker: HashCounter<String>,
     pub label: HashCounter<String>,
     pub series: HashCounter<String>,
-    pub duration: HashCounter<usize>,
+    pub duration: HashCounter<u32>,
 }
 
 impl IdentTable {
@@ -80,7 +80,7 @@ enum IdentKind {
     Speaker(String),
     Label(String),
     Series(String),
-    Duration(usize),
+    Duration(u32),
 }
 
 fn ident_of<F>(sentence: &Sentence, mut f: F)
